requirements.txt
click==8.1.7
PyYAML==6.0.1
boto3==1.34.144
  botocore==1.34.144
    python-dateutil==2.9.0.post0
    urllib3==2.2.2
  jmespath==1.0.1
  s3transfer==0.10.2
docker==7.1.0
  requests==2.32.3
    # urllib3
    certifi==2024.7.4
    charset-normalizer==3.3.2
    idna==3.7
  # urllib3
build==1.2.1
  packaging==24.1
  pyproject_hooks==1.1.0
  tomli==2.0.1
